mejora caja movimientos

+++ backend/types/ventas.go
@@ -9,14 +9,14 @@ type Caja struct {
 	Nombre       string `db:"nombre"`
 	Descripcion  string `db:"descripcion"`
 	MonedaTipo   int8   `db:"moneda"`
-	CuadreFecha  int64  `db:"cuadre_fecha"`
+	CuadreFecha  int32  `db:"cuadre_fecha"`
 	CuadreSaldo  int32  `db:"cuadre_saldo"`
 	SaldoCurrent int32  `db:"saldo_current"`
 	// Propiedades generales
 	Status    int8  `json:"ss,omitempty" db:"status,view"`
-	Updated   int64 `json:"upd,omitempty" db:"updated,view"`
+	Updated   int32 `json:"upd,omitempty" db:"updated,view"`
 	UpdatedBy int32 `json:",omitempty" db:"updated_by"`
-	Created   int64 `json:",omitempty" db:"created"`
+	Created   int32 `json:",omitempty" db:"created"`
 	CreatedBy int32 `json:",omitempty" db:"created_by"`
 }
 
@@ -30,10 +30,11 @@ type CajaMovimiento struct {
 	Tipo       int8  `json:",omitempty" db:"tipo"`
 	SaldoFinal int32 `db:"saldo_final"`
 	Monto      int32 `db:"monto"`
-	Created    int64 `json:",omitempty" db:"created"`
+	Created    int32 `json:",omitempty" db:"created"`
 	CreatedBy  int32 `json:",omitempty" db:"created_by,view"`
 }
 
+/*
 func (e *CajaMovimiento) MakeID(unixTimeMill int64) int64 {
 	return int64(e.CajaID)*10_000_000_000_000 + unixTimeMill
 }
@@ -41,6 +42,7 @@ func (e *CajaMovimiento) MakeID(unixTimeMill int64) int64 {
 func (e *CajaMovimiento) SetID(unixTimeMill int64) {
 	e.ID = int64(e.CajaID)*10_000_000_000_000 + unixTimeMill
 }
+*/
 
 type CajaCuadre struct {
 	TAGS            `table:"caja_cuadre"`
@@ -52,10 +54,12 @@ type CajaCuadre struct {
 	SaldoSistema    int32 `json:",omitempty" db:"saldo_sistema"`
 	SaldoReal       int32 `json:",omitempty" db:"saldo_real"`
 	SaldoDiferencia int32 `json:",omitempty" db:"saldo_diferencia"`
-	Created         int64 `json:",omitempty" db:"created"`
+	Created         int32 `json:",omitempty" db:"created"`
 	CreatedBy       int32 `json:",omitempty" db:"created_by,view"`
 }
 
+/*
 func (e *CajaCuadre) SetID(unixTimeMill int64) {
 	e.ID = int64(e.CajaID)*10_000_000_000_00 + unixTimeMill
 }
+*/
